separate table of contents into its own generic component
