+++ client/src/components/common/TableOfContents/TableOfContents.types.ts
@@ -1,0 +1,4 @@
+export interface TOCHeader {
+  id: string;
+  text: string;
+}
