+++ client/src/components/common/TableOfContents/index.ts
@@ -1,0 +1,1 @@
+export * from './TableOfContents';
