+++ client/src/components/common/index.ts
@@ -5,3 +5,4 @@ export * from './ErrorMessage';
 export * from './Link';
 export * from './Markdown';
 export * from './Overlay';
+export * from './TableOfContents';
